pyproject.toml
[project]
name = "recipes"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.64.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
